+++ go.mod
@@ -1,3 +1,5 @@
 module todo
 
 go 1.22.4
+
+require github.com/joho/godotenv v1.5.1
