+++ index.html
@@ -9,6 +9,7 @@
       integrity="sha384-Y7hw+L/jvKeWIRRkqWYfPcvVxHzVzn5REgzbawhxAuQGwX1XWe70vji+VSeHOThJ"
       crossorigin="anonymous"
     ></script>
+    <link rel="stylesheet" href="./static/css/main.css" />
   </head>
   <body>
     <h1>Todo List:</h1>
@@ -27,6 +28,7 @@
       hx-delete="/delete-todos/"
       hx-target="#todo-wrapper"
       hx-swap="innerHTML"
+      id="delete"
     >
       Delete
     </button>
