+++ static/css/main.css
@@ -1,0 +1,5 @@
+body {
+  background-color: orange;
+  color: blue;
+  font: 200;
+}
